Move static method to abstract class
