+++ src/main/java/io/github/townyadvanced/townyprovinces/jobs/map_display/DisplayProvincesOnDynmapAction.java
@@ -220,19 +220,6 @@ public class DisplayProvincesOnDynmapAction extends DisplayProvincesOnMapAction 
 		return result;
 	}
 
-	/**
-	 * Find the border coords around the given province
-	 * 
-	 * Note that these co-cords will not actually belong to the province
-	 */
-	public static Set<TPCoord> findAllBorderCoords(Province province) {
-		Set<TPCoord> resultSet = new HashSet<>();
-		for(TPCoord provinceCoord: province.getListOfCoordsInProvince()) {
-			resultSet.addAll(province.getAdjacentBorderCoords(provinceCoord));
-		}
-		return resultSet;
-	}
-
 	private void drawBorderLine(List<TPCoord> drawableLineOfBorderCoords, Province province, String markerId) {
 		String worldName = TownyProvincesSettings.getWorldName();
 		int landBorderColour = TownyProvincesSettings.getLandProvinceBorderColour();
