+++ src/main/java/io/github/townyadvanced/townyprovinces/jobs/map_display/DisplayProvincesOnPl3xMapV3Action.java
@@ -259,19 +259,6 @@ public class DisplayProvincesOnPl3xMapV3Action extends DisplayProvincesOnMapActi
 		return result;
 	}
 
-	/**
-	 * Find the border coords around the given province
-	 * 
-	 * Note that these co-cords will not actually belong to the province
-	 */
-	public static Set<TPCoord> findAllBorderCoords(Province province) {
-		Set<TPCoord> resultSet = new HashSet<>();
-		for(TPCoord provinceCoord: province.getListOfCoordsInProvince()) {
-			resultSet.addAll(province.getAdjacentBorderCoords(provinceCoord));
-		}
-		return resultSet;
-	}
-
 	private void drawBorderLine(List<TPCoord> drawableLineOfBorderCoords, Province province, String markerId) {
 		int landBorderColour = TownyProvincesSettings.getLandProvinceBorderColour() +
 			(int)(255*TownyProvincesSettings.getLandProvinceBorderOpacity()) << 24;
